feat(pages): register page finished
