+++ src/Components/Brand/Brand.module.css
@@ -2,6 +2,7 @@
     font-family: 'Quicksand';
     color: #3564FD;
     font-weight: 600;
+    margin: 0;
 }
 .sm{
     font-size: 1.2rem;
