+++ src/Components/Forms/UserForm/UserRegistrationForm.jsx
@@ -1,39 +1,38 @@
-import React from 'react';
-import * as Yup from 'yup';
-import UserForm from './UserForm';
+import React from "react";
+import * as Yup from "yup";
+import UserForm from "./UserForm";
 
-export default function  UserRegistrationForm () {
+export default function UserRegistrationForm() {
   return (
-    <div className="container">
-      <h1>Sign Up</h1>
-      <UserForm
-        initialValues={{
-          firstName: '',
-          lastName: '',
-          email: '',
-          password: '',
-          confirmPassword: ''
-        }}
-        validationSchema={Yup.object({
-          firstName: Yup.string()
-            .max(15, 'Must be 15 characters or less')
-            .required('Required'),
-          lastName: Yup.string()
-            .max(20, 'Must be 20 characters or less')
-            .required('Required'),
-          email: Yup.string()
-            .email('Invalid email address')
-            .required('Required'),
-          password: Yup.string()
-            .min(8, 'Password must be at least 8 characters')
-            .required('Required'),
-          confirmPassword: Yup.string()
-            .oneOf([Yup.ref('password'), null], 'Passwords must match')
-            .required('Required')
-        })}
-        action="register"
-      >
-      </UserForm>
-    </div>
+    <UserForm
+      initialValues={{
+        firstName: "",
+        lastName: "",
+        email: "",
+        password: "",
+        confirmPassword: "",
+      }}
+      validationSchema={Yup.object({
+        firstName: Yup.string()
+          .max(15, "Necesita tener como máximo 15 caracteres.")
+          .required("Este campo es requerido."),
+        lastName: Yup.string()
+          .max(20, "Necesita tener como máximo 20 caracteres.")
+          .required("Este campo es requerido."),
+        email: Yup.string()
+          .email("Email invalido.")
+          .required("Este campo es requerido."),
+        password: Yup.string()
+          .min(8, "La contraseña necesita tener como mínimo 8 caracteres.")
+          .required("Este campo es requerido."),
+        confirmPassword: Yup.string()
+          .oneOf(
+            [Yup.ref("password"), null],
+            "Las contraseñas deben ser iguales"
+          )
+          .required("Este campo es requerido."),
+      })}
+      action="register"
+    />
   );
-};
+}
