+++ src/Components/Layouts/FormsLayout/FormsLayout.module.css
@@ -1,18 +1,18 @@
 .formWrapper{
-    min-height: 100vh;
-    position: relative;
+    min-height: 86vh;
+    padding: 0;
 }
 .decorationBackground{
     position: relative;
     background-color: #f9f9ff;
-    height: 100vh;
+    min-height: 100vh;
     display: flex;
 }
 .iconWrapper{
     display: flex;
     width: 100%;
     background-color: #f9f9ff;
-    height: calc(100vh - 6rem);
+    height: 100%;
     justify-content: center;
     align-items: center;
 }
