+++ src/Components/Layouts/FormsLayout/components/FormsFooter/components/SecureFooter/SecureFooter.module.css
@@ -1,8 +1,5 @@
 .footer{
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    right: 0;
+    width: 100%;
     padding: 1.3rem 2rem;
     display: flex;
     justify-content: center;
