+++ src/pages/Login/Login.jsx
@@ -1,273 +1,29 @@
 import React from "react";
 import LoginForm from "../../Components/Forms/LoginForm/LoginForm";
 import Brand from "../../Components/Brand/Brand";
-import LoginFooter from "./components/LoginFooter/LoginFooter";
-import styles from "./Login.module.css";
-import AppStoresFooter from "./components/AppStoresFooter/AppStoresFooter";
+import FormsLayout from "../../Components/Layouts/FormsLayout/FormsLayout";
+import Button from "../../Components/Button/Button";
+import { useNavigate } from "react-router-dom";
 
 export default function Login() {
+  const navigate = useNavigate();
+  const handleRedirect = link => navigate(link);
   return (
-    <div className="container-fluid">
-      <div className="row">
-        <div
-          className={` ${styles.formWrapper} col-12 col-md-6 h-100 d-flex flex-column align-items-center`}
-        >
-          <div className="text-center mb-5 mt-5">
-            <Brand size="xl" />
-          </div>
-          <div className="col-8">
-            <LoginForm />
-          </div>
-          <LoginFooter />
-        </div>
-        <div
-          className={`${styles.decorationBackground}col-12  col-md-6 d-none d-md-flex position-relative p-0`}
-        >
-          <div className={`${styles.iconWrapper} `}>
-            <svg
-              id="login-illustration"
-              width="360"
-              height="360"
-              viewBox="0 0 360 360"
-              fill="none"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <g clip-path="url(#clip0_805:34271)">
-                <path
-                  d="M29.3379 119.594C25.3143 149.294 46.7555 208.641 74.3379 174.6C101.92 140.559 181.279 85.8176 128.073 65.2764C110.973 58.7117 35.1084 76.6588 29.3379 119.594Z"
-                  fill="#E7EEFF"
-                ></path>
-                <path
-                  d="M26.0551 309.335C26.1609 312.3 26.4256 315.265 26.9021 318.229C31.2433 344.594 73.3315 384.3 201.396 340.888C295.579 308.912 341.161 225.847 338.673 147.494C337.773 119.276 330.626 91.694 317.284 67.2881C269.373 -20.3825 183.608 14.8763 185.461 72.3175C185.567 75.2822 185.884 78.2999 186.467 81.4234C188.637 92.9646 190.12 103.87 190.437 114.141C191.867 159.459 170.902 193.129 88.6845 222.459C57.2374 233.735 24.7845 270.053 26.0551 309.335Z"
-                  fill="#E7EEFF"
-                ></path>
-                <path
-                  d="M11.3903 236.859C12.9256 226.641 21.1315 218.329 31.3492 216.688C37.0139 215.788 42.3609 216.847 46.808 219.282C47.655 208.218 57.1845 199.588 68.6198 200.382C76.7727 200.912 83.8668 206.576 86.5139 214.306C87.9433 218.541 87.8374 222.512 86.8315 226.165L86.7786 274.129C86.7786 289.376 74.3903 301.765 59.0903 301.765C43.7903 301.765 31.4021 289.376 31.4021 274.076V264.441C18.7492 262.429 9.37858 250.518 11.3903 236.859Z"
-                  stroke="#F088FF"
-                  stroke-width="1.81884"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M117.591 37.906H200.708C217.491 37.906 231.096 51.5119 231.096 68.2942V274.765H87.2023V68.2942C87.2023 51.5119 100.808 37.906 117.591 37.906Z"
-                  fill="#FAFAFA"
-                  stroke="#3564FD"
-                  stroke-width="2.72826"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M87.2027 274.765V81.7413C84.9709 58.7648 91.7027 37.906 121.932 37.906H120.238C139.403 37.906 154.967 57.5472 154.967 81.7413V250.571L87.2027 274.765Z"
-                  fill="#3E6BFD"
-                ></path>
-                <path
-                  d="M191.656 69.2467H185.991C185.674 69.2467 185.356 69.4585 185.25 69.7762L177.044 90.5291C176.938 90.7938 176.991 91.0585 177.15 91.2703C177.203 91.3761 177.256 91.4291 177.362 91.482C177.521 91.5879 177.68 91.6409 177.838 91.6409H209.55C209.815 91.6409 210.08 91.535 210.238 91.2703C210.397 91.0585 210.397 90.7938 210.344 90.5291L202.086 69.7762C201.98 69.4585 201.662 69.2467 201.344 69.2467H193.403"
-                  fill="#3564FD"
-                ></path>
-                <path
-                  d="M231.097 274.765V294.088C231.097 295.412 230.038 296.418 228.767 296.418H208.385C207.22 296.418 206.32 297.318 206.32 298.482V313.783C206.32 314.841 205.42 315.741 204.361 315.741H48.1849V303.565C48.1849 299.753 51.3085 296.63 55.1202 296.63H71.5849V282.071C71.5849 278.259 74.7085 275.135 78.5202 275.135H87.2026"
-                  fill="#F088FF"
-                ></path>
-                <path
-                  d="M212.515 296.418H70.8442"
-                  stroke="#FAFAFA"
-                  stroke-width="0.90942"
-                  stroke-miterlimit="10"
-                ></path>
-                <path
-                  d="M193.297 39.3883V74.7"
-                  stroke="#3564FD"
-                  stroke-width="1.81884"
-                  stroke-miterlimit="10"
-                ></path>
-                <path
-                  d="M349.677 93.8647C345.812 93.1765 342.053 94.553 339.512 97.2C336.494 90.1589 329.506 85.2883 321.406 85.2883C314.259 85.2883 308.065 89.1 304.571 94.7647C302.136 92.5941 298.906 91.2706 295.412 91.2706C287.789 91.2706 281.647 97.4647 281.647 105.035C281.647 116.629 295.412 116.471 295.412 116.471C295.412 116.471 340.624 116.735 344.594 116.947C350.841 117.212 357.83 113.135 358.889 106.994C359.947 100.853 355.818 94.9765 349.677 93.8647Z"
-                  stroke="#F088FF"
-                  stroke-width="1.81884"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M108.741 7.51779C105.83 6.98838 103.024 8.0472 101.118 10.006C98.8415 4.76485 93.6003 1.05897 87.5121 1.05897C82.165 1.05897 77.5062 3.91779 74.9121 8.15309C73.0591 6.51191 70.6768 5.50603 68.0297 5.50603C62.3121 5.50603 57.7062 10.1119 57.7062 15.8296C57.7062 24.5119 68.0297 24.406 68.0297 24.406C68.0297 24.406 101.912 24.6178 104.877 24.7766C109.536 24.9884 114.777 21.9178 115.571 17.3119C116.418 12.706 113.347 8.31191 108.741 7.51779Z"
-                  stroke="#F088FF"
-                  stroke-width="1.81884"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M230.302 256.235H154.967L123.467 268.571H198.802L230.302 256.235Z"
-                  fill="#3564FD"
-                ></path>
-                <path
-                  d="M229.084 223.359L208.914 223.094L209.179 187.888L208.967 187.994C208.967 187.941 208.967 187.941 208.967 187.888C209.02 179.365 202.19 172.429 193.667 172.376C185.143 172.323 178.208 179.153 178.155 187.676C178.102 192.971 180.749 197.682 184.826 200.435L184.561 239.506L231.043 239.929"
-                  stroke="#3564FD"
-                  stroke-width="2.72826"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M231.149 211.553L214.896 211.341C211.667 211.341 208.967 213.935 208.967 217.218C208.967 220.447 211.561 223.147 214.843 223.147L231.096 223.359"
-                  stroke="#3564FD"
-                  stroke-width="2.72826"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M230.99 168.988L230.355 168.935C218.761 168.882 209.337 178.359 209.549 189.953L209.814 206.206C209.814 209.435 212.408 212.135 215.69 212.135L231.255 212.294"
-                  stroke="#3564FD"
-                  stroke-width="2.72826"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <path
-                  d="M194.515 193.227C197.609 193.227 200.118 190.718 200.118 187.623C200.118 184.529 197.609 182.02 194.515 182.02C191.42 182.02 188.911 184.529 188.911 187.623C188.911 190.718 191.42 193.227 194.515 193.227Z"
-                  fill="#F088FF"
-                ></path>
-                <path
-                  d="M156.026 161.047C153.432 161.047 151.261 161.894 150.203 164.859H144.008V174.229H150.097C151.05 177.406 153.326 178.412 156.026 178.412C159.573 178.412 162.432 174.494 162.432 169.729C162.432 164.965 159.573 161.047 156.026 161.047Z"
-                  fill="#FAFAFA"
-                  stroke="#3E6BFD"
-                  stroke-width="1.81884"
-                  stroke-miterlimit="10"
-                  stroke-linecap="round"
-                  stroke-linejoin="round"
-                ></path>
-                <mask
-                  id="mask0_805:34271"
-                  maskUnits="userSpaceOnUse"
-                  x="86"
-                  y="37"
-                  width="69"
-                  height="238"
-                  style={{ "mask-type": "alpha" }}
-                >
-                  <path
-                    d="M87.216 274.765V81.7413C84.9842 58.7648 91.716 37.906 121.945 37.906H120.251C139.416 37.906 154.981 57.5472 154.981 81.7413V250.571L87.216 274.765Z"
-                    fill="#3E6BFD"
-                  ></path>
-                </mask>
-                <g mask="url(#mask0_805:34271)">
-                  <mask
-                    id="mask1_805:34271"
-                    maskUnits="userSpaceOnUse"
-                    x="-42"
-                    y="2"
-                    width="325"
-                    height="324"
-                    style={{ "mask-type": "alpha" }}
-                  >
-                    <rect
-                      x="-41.2933"
-                      y="2.11769"
-                      width="323.471"
-                      height="323.471"
-                      fill="#C4C4C4"
-                    ></rect>
-                  </mask>
-                  <g mask="url(#mask1_805:34271)">
-                    <path
-                      d="M125.476 100.507C114.694 94.4417 108.629 99.8329 104.586 101.855C100.542 103.876 95.825 103.202 93.8033 100.507C91.7816 97.8112 93.8033 93.7678 98.5206 89.0505C102.294 85.2767 114.02 64.1163 128.846 54.0079C149.063 42.5516 139.519 31.0332 154.454 26.3781C206.344 10.2046 177.367 -16.0774 163.889 -37.6421"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M29.7815 171.94C20.3469 188.113 3.22992 184.339 -0.543901 190.809C-5.26118 198.896 15.6296 190.539 15.6296 211.7C15.6296 227.873 28.1017 213.545 32.4771 206.982C45.955 186.765 14.2818 189.461 52.694 167.223C65.6433 159.726 74.9326 152.397 78.9759 142.288C83.0193 132.18 93.1278 102.528 112.671 111.963C132.214 121.398 90.4322 155.092 86.3888 159.136C82.3454 163.179 79.6738 169.591 79.6498 175.309C79.2006 180.476 74.2587 192.426 58.0851 198.896C37.8682 206.982 31.7037 219.859 32.4771 233.264C34.4987 268.307 3.49948 282.459 -14.6957 287.176C-32.891 291.893 -42.9994 314.806 -28.8476 325.588"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M174.67 334.349C181.614 340.627 195.561 287.85 222.146 321.545C250.893 357.979 269.015 307.393 224.796 286.178C202.889 275.668 203.909 300.383 190.041 258.22C185.366 244.006 180.103 233.409 171.042 227.374C161.98 221.339 135.04 205.352 148.29 188.165C161.54 170.979 185.927 218.794 189.053 223.582C192.179 228.371 197.905 232.303 203.496 233.502C208.46 235.004 219.139 242.297 222.146 259.455C225.904 280.902 237.239 289.582 250.517 291.58C285.227 296.805 292.704 330.051 293.581 348.828C294.457 367.604 314.803 382.206 328.264 370.573"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M207.3 -39.6638C205.278 -35.6204 198.944 -26.5901 195.17 -20.1207C190.453 -12.034 187.083 -3.94719 193.148 0.0961905C199.213 4.13957 207.3 -5.29499 209.996 -12.7079C212.691 -20.1207 218.083 -35.6204 230.213 -44.381C242.343 -53.1417 252.451 -59.2067 256.495 -69.3152C260.538 -79.4237 270.646 -109.075 290.189 -99.6406C309.733 -90.206 267.951 -56.5112 263.908 -52.4678C259.864 -48.4244 257.192 -42.0125 257.169 -36.2943C256.719 -31.1277 251.777 -19.1772 235.604 -12.7079C215.387 -4.62116 209.222 8.25581 209.996 21.6608C212.017 56.7035 181.018 70.8553 162.823 75.5726C144.628 80.2899 130.191 99.8329 144.343 110.615C158.495 120.724 191.516 116.006 191.516 132.854C191.516 139.89 182.755 146.332 166.581 147.68"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M125.476 100.507C132.889 104.55 144.345 115.333 132.889 137.571C122.682 157.384 103.687 165.65 96.4986 168.57C88.6364 172.165 75.0686 183.531 83.6945 200.244C94.4769 221.134 107.955 226.526 92.4552 238.656C76.9555 250.786 62.9633 253.673 54.7169 266.285C43.2607 283.807 77.1822 295.899 72.2383 279.09C68.8688 267.633 72.8341 252.957 83.6945 254.829C103.238 258.199 85.0423 283.133 78.3033 296.611C71.5644 310.089 78.9772 325.588 95.1508 326.936"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M166.582 147.68C149.735 149.701 130.192 155.766 114.692 175.309C102.292 190.944 111.323 205.185 117.388 210.352C121.431 215.743 126.822 229.895 114.018 232.591C94.9969 236.595 107.953 249.438 100.54 260.22C92.2288 275.495 83.4233 311.571 114.692 333.675"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M15.6301 327.61C33.1514 328.284 26.4125 314.132 29.108 308.067C39.2165 293.241 35.847 324.241 44.6077 323.567C53.3683 322.893 52.0205 312.11 46.6294 299.98C41.2382 287.85 31.8037 285.155 25.0647 298.903C20.5431 308.127 -1.21732 330.306 -18.0647 322.893C-36.9338 314.59 -26.1515 298.903 -14.0213 303.35C-7.44109 305.762 3.82085 302.586 7.54337 299.98C25.8368 287.176 24.0794 281.266 33.9474 269.701C47.3397 254.006 47.0219 203.426 75.2269 212.669C99.1585 220.512 78.7124 238.819 57.4117 254.155C40.5643 266.285 37.8687 282.85 52.0206 286.502C72.9114 291.893 73.3157 316.423 56.064 328.284"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M283.841 176.657C271.037 145.658 237.342 129.484 211.467 125.441C190.767 122.206 199.616 104.775 206.627 96.4632C219.5 81.2049 213.366 112.637 222.127 111.963C230.888 111.289 229.54 100.507 224.149 88.3764C218.758 76.2463 209.323 73.5507 202.584 87.2989C198.062 96.5233 176.302 118.702 159.455 111.289C140.586 102.986 151.368 87.2988 163.498 91.7459C170.078 94.1583 181.34 90.9819 185.063 88.3764C203.356 75.5723 201.599 69.6619 211.467 58.0972C224.859 42.4024 224.541 -8.17823 252.746 1.06517C276.678 8.90805 256.232 27.215 234.931 42.5514C218.084 54.6816 209.997 73.268 224.149 76.9201C230.97 78.6805 250.577 78.1469 244.755 105.898C238.933 133.648 313.492 172.613 304.057 120.049"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                      stroke-linecap="round"
-                    ></path>
-                    <path
-                      d="M122.593 335.386C124.315 327.12 126.983 310.458 122.96 298.545C117.931 283.654 98.2199 252.739 117.041 243.089C134.223 234.28 138.368 276.264 137.52 283.503C136.671 290.743 135.892 322.134 143.669 329.632C152.194 337.743 183.076 330.306 190.168 298.545C193.527 283.503 188.955 274.911 167.929 252.808C141.647 225.178 110.648 175.983 153.778 165.201C196.907 154.419 207.294 118.107 235.993 145.658C252.841 161.831 243.406 209.678 288.557 198.222"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                    <path
-                      d="M29.7817 171.94C40.5641 140.266 52.1551 156.844 63.4766 147.679C77.6284 136.223 64.8244 103.876 83.6935 87.0286C102.563 70.1812 107.28 72.8768 111.323 62.0944C112.671 58.9496 117.989 45.3453 122.106 39.8558C134.236 23.6823 169.952 20.3128 165.909 -3.27357"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                    <path
-                      d="M180.52 289.511C182.397 303.019 169.626 333.675 153.779 330.305C138.142 324.779 156.128 296.265 149.873 285.827C143.618 275.388 146.807 261.88 156.128 261.88C165.909 261.88 178.644 276.002 180.52 289.511Z"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                    <path
-                      d="M-58.4996 99.8328C-47.9419 98.0357 -22.3023 86.892 2.82497 95.1155C39.8893 107.246 27.7592 134.875 38.5415 140.267C49.3239 145.658 65.4974 137.571 58.0845 122.745C43.9327 99.8328 59.74 84.4255 70.2147 77.5942C85.7143 67.4858 111.322 65.4641 118.061 33.117C123.452 7.23936 108.627 -3.72269 100.54 -5.96902"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                    <path
-                      d="M89.7576 -3.94738C98.5183 3.46549 117.657 25.7041 105.257 41.8776C89.7576 62.0945 70.8885 74.8986 62.8018 74.8986C54.659 74.8986 35.172 64.7901 50.6716 46.5949C66.1712 28.3997 74.258 4.13938 60.1062 -3.94738"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                    <path
-                      d="M66.8573 -5.58425C74.8136 -2.73364 82.8469 12.3185 77.7263 22.1761C68.7138 39.5259 50.0644 50.6263 60.7496 62.8187C66.2816 69.131 93.2135 48.7008 96.8404 37.4322C100.467 26.1635 103.545 2.48036 74.9048 -6.3794"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                    <path
-                      d="M190.169 -9.33858C192.865 -0.128653 193.404 20.852 173.996 31.0952C149.735 43.8993 119.41 74.8985 128.845 82.3114C138.279 89.7243 151.757 54.6816 169.278 51.986C186.8 49.2905 206.343 16.9434 202.973 -3.27351"
-                      stroke="#FFE2E2"
-                      stroke-width="1.15762"
-                    ></path>
-                  </g>
-                </g>
-              </g>
-              <defs>
-                <clipPath id="clip0_805:34271">
-                  <rect width="360" height="360" fill="white"></rect>
-                </clipPath>
-              </defs>
-            </svg>
-          </div>
-          <AppStoresFooter />
-        </div>
+    <FormsLayout>
+      <div className="text-center mb-5 mt-5">
+        <Brand size="xl" />
       </div>
-    </div>
+      <div className="col-10 col-md-6">
+        <LoginForm />
+        <Button
+          variant="outlined"
+          size="md"
+          fullWidth
+          handleClick={() => handleRedirect("/register")}
+        >
+          Registrarse
+        </Button>
+      </div>
+    </FormsLayout>
   );
 }
